refactor(scripts): create dev directory
